+++ src/blueprints/__init__.py
@@ -1,1 +1,2 @@
 from .telegram_bot import telegram_bot_blueprint
+from .legal import legal_blueprint