+++ src/blueprints/legal/views.py
@@ -1,0 +1,23 @@
+from flask import redirect, url_for
+
+from src.blueprints.legal import legal_blueprint as bp
+
+
+@bp.route("/privacy-policy")
+def privacy_policy():
+    return redirect(
+        url_for(
+            "static",
+            filename="legal/privacy_policy.txt"
+        )
+    )
+
+
+@bp.route("/terms-and-conditions")
+def terms_and_conditions():
+    return redirect(
+        url_for(
+            "static",
+            filename="legal/terms_and_conditions.txt"
+        )
+    )